+++ NEWS.md
@@ -1,5 +1,11 @@
 # MetaboAnnotation 0.99
 
+## Changes in 0.99.5
+
+- Definitions of "`score`", "`score_rt`" changed to be the difference
+  (with sign) between target and query m/z or retention time respectively.
+- `"ppm_error"` becomes error without sign.
+
 ## Changes in 0.99.4
 
 - Add matches m/z error (variable `"ppm_error"`) to the `Matched` object